Git merge resolved
The merge issues have been resolved, and right now Nik_Branch is the
trunk. The name will be changed, if possible. Curtis_Branch will not be
used for simplicity.

There are still bugs/errors that have come during the development
process, and our first objective in the next release is to resolve
those.